Condense videos-to-process storage
